requirements.txt
flet
peewee
psycopg2-binary
python-dotenv